Use CartesianIndices of Slice{OneTo} instead of OneTo
This corresponds more closely with how the normal call chain/compiler
lowers array access syntax with colons.
For instance
```julia
    A[:,:] = x
```
is lowered to
```julia
    setindex!(A, x, :, :)
```
The `:`s are converted to actual indices in an internal function by
calling `to_indices()`, so the operation is more like
```julia
    setindex!(A, x, to_indices(A, :, :)...)
```
The indices are `(:, :)` → `axes(A)`, but to retain the fact that the
indices span the entire appropriate dimension of `A`, the returned
axes are wrapped in `Base.Slice` types (which correspondingly have
trivial bounds checking logic).

We know we just asked for the `axes(x)`, so just map each axis to a
`Slice`. Alternatively, the same could be achieved with
```julia
    I = CartesianIndices(to_indices(x, ntuple(_ -> :, ndims(x))))
```
if we insisted on relying on `to_indices` rather than explicitly
making use of `Base.Slice`.

This change has the added benefit of slightly simplifying the Julia
AST/IR, and --- when bounds-checking is enforced --- the LLVM IR as well.
(Normally the entire `_legendre_impl!` function is `at-inbounds`
within `_legendre!`, so the bounds checking is skipped eliminated as
dead code.)