Add new 4π sphere (geodesy) normalization
Defined such that the integrating square of this normalization times the complex exponential
in azimuth equals the surface area of the unit sphere. This complements the spherical
normalization `LegendreSphereNorm` which instead integrates over the sphere to unity.

Also adds an intermediate `AbstractLegendreSphereNorm <: AbstractLegendreNorm`
in the abstract type hierarchy to consolidate the coefficients for
all three of `LegendreFourPiNorm`, `LegendreOrthoNorm`, and
`LegendreSphereNorm`.

The original motivation was to see if replacing the initial condition of 1 / sqrt(4π) which
must necessarily be rounded (because π is trancendental) with an exactly-representable value
(1.0 for the 4π normalization) could lead to better numerical precision. Namely, I thought
that something like summing over many terms could show the accumulated bias, but I haven't
actually found that to be true in practice. Nonetheless, having another option for a
built-in normalizations is still useful.